requirements.txt
Logbook~=1.5.3
RPi.GPIO~=0.7.0
CASlibrary~=0.6.0